third_party/blink/web_tests/fast/dom/shadow/element-name-in-shadow-expected.txt

CONSOLE WARNING: line 11: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
<a>: Before appendChild, named_a should not be in document.
PASS element.name in document is false
PASS element.name in window is false
<a>: After appendChild, named_a should not still be in document, because the element was inserted into a shadow DOM tree.
PASS element.name in document is false
PASS element.name in window is false
<a>: After changing its name attribute into renamed_a by using setAttribute, the name should not still be in document.
PASS element.name in document is false
PASS element.name in window is false
<a>: After changing its name attribute into re-renamed_a by using '.name=', the name should not still be in document.
PASS element.name in document is false
PASS element.name in window is false
<area>: Before appendChild, named_area should not be in document.
PASS element.name in document is false
PASS element.name in window is false
<area>: After appendChild, named_area should not still be in document, because the element was inserted into a shadow DOM tree.
PASS element.name in document is false
PASS element.name in window is false
<area>: After changing its name attribute into named_area by using setAttribute, the name should not still be in document.
PASS element.name in document is false
PASS element.name in window is false
<area>: After changing its name attribute into re-renamed_area by using '.name=', the name should not still be in document.
PASS element.name in document is false
PASS element.name in window is false
<embed>: Before appendChild, named_embed should not be in document.
PASS element.name in document is false
PASS element.name in window is false
<embed>: After appendChild, named_embed should not still be in document, because the element was inserted into a shadow DOM tree.
PASS element.name in document is false
PASS element.name in window is false
<embed>: After changing its name attribute into renamed_embed by using setAttribute, the name should not still be in document.
PASS element.name in document is false
PASS element.name in window is false
<embed>: After changing its name attribute into re-renamed_embed by using '.name=', the name should not still be in document.
PASS element.name in document is false
PASS element.name in window is false
<form>: Before appendChild, named_form should not be in document.
PASS element.name in document is false
PASS element.name in window is false
<form>: After appendChild, named_form should not still be in document, because the element was inserted into a shadow DOM tree.
PASS element.name in document is false
PASS element.name in window is false
<form>: After changing its name attribute into renamed_form by using setAttribute, the name should not still be in document.
PASS element.name in document is false
PASS element.name in window is false
<form>: After changing its name attribute into re-renamed_form by using '.name=', the name should not still be in document.
PASS element.name in document is false
PASS element.name in window is false
<frame>: Before appendChild, named_frame should not be in document.
PASS element.name in document is false
PASS element.name in window is false
<frame>: After appendChild, named_frame should not still be in document, because the element was inserted into a shadow DOM tree.
PASS element.name in document is false
PASS element.name in window is false
<frame>: After changing its name attribute into renamed_frame by using setAttribute, the name should not still be in document.
PASS element.name in document is false
PASS element.name in window is false
<frame>: After changing its name attribute into re-renamed_frame by using '.name=', the name should not still be in document.
PASS element.name in document is false
PASS element.name in window is false
<frameset>: Before appendChild, named_frameset should not be in document.
PASS element.name in document is false
PASS element.name in window is false
<frameset>: After appendChild, named_frameset should not still be in document, because the element was inserted into a shadow DOM tree.
PASS element.name in document is false
PASS element.name in window is false
<frameset>: After changing its name attribute into named_frameset by using setAttribute, the name should not still be in document.
PASS element.name in document is false
PASS element.name in window is false
<frameset>: After changing its name attribute into re-renamed_frameset by using '.name=', the name should not still be in document.
PASS element.name in document is false
PASS element.name in window is false
<iframe>: Before appendChild, named_iframe should not be in document.
PASS element.name in document is false
PASS element.name in window is false
<iframe>: After appendChild, named_iframe should not still be in document, because the element was inserted into a shadow DOM tree.
PASS element.name in document is false
PASS element.name in window is false
<iframe>: After changing its name attribute into renamed_iframe by using setAttribute, the name should not still be in document.
PASS element.name in document is false
PASS element.name in window is false
<iframe>: After changing its name attribute into re-renamed_iframe by using '.name=', the name should not still be in document.
PASS element.name in document is false
PASS element.name in window is false
<img>: Before appendChild, named_img should not be in document.
PASS element.name in document is false
PASS element.name in window is false
<img>: After appendChild, named_img should not still be in document, because the element was inserted into a shadow DOM tree.
PASS element.name in document is false
PASS element.name in window is false
<img>: After changing its name attribute into renamed_img by using setAttribute, the name should not still be in document.
PASS element.name in document is false
PASS element.name in window is false
<img>: After changing its name attribute into re-renamed_img by using '.name=', the name should not still be in document.
PASS element.name in document is false
PASS element.name in window is false
<object>: Before appendChild, named_object should not be in document.
PASS element.name in document is false
PASS element.name in window is false
<object>: After appendChild, named_object should not still be in document, because the element was inserted into a shadow DOM tree.
PASS element.name in document is false
PASS element.name in window is false
<object>: After changing its name attribute into renamed_object by using setAttribute, the name should not still be in document.
PASS element.name in document is false
PASS element.name in window is false
<object>: After changing its name attribute into re-renamed_object by using '.name=', the name should not still be in document.
PASS element.name in document is false
PASS element.name in window is false
PASS successfullyParsed is true

TEST COMPLETE
